+++ packages/@ionic/cli-framework/src/lib/__tests__/validators.ts
@@ -1,4 +1,4 @@
-import { stripAnsi } from '@ionic/cli-framework-output';
+import { stripAnsi } from '@ionic/utils-terminal';
 
 import { contains, validators } from '../validators';