+++ packages/@ionic/cli/src/commands/cordova/base.ts
@@ -1,7 +1,7 @@
 import { MetadataGroup } from '@ionic/cli-framework';
-import { prettyPath } from '@ionic/cli-framework/utils/format';
 import { mkdirp, pathExists } from '@ionic/utils-fs';
 import { ERROR_COMMAND_NOT_FOUND, ERROR_SIGNAL_EXIT, SubprocessError } from '@ionic/utils-subprocess';
+import { prettyPath } from '@ionic/utils-terminal';
 import * as lodash from 'lodash';
 import * as path from 'path';