+++ packages/@ionic/cli/src/commands/package/deploy.ts
@@ -6,8 +6,8 @@ import {
   validators,
 } from '@ionic/cli-framework';
 import { LOGGER_LEVELS } from '@ionic/cli-framework-output';
-import { columnar } from '@ionic/cli-framework/utils/format';
 import { sleep } from '@ionic/utils-process';
+import { columnar } from '@ionic/utils-terminal';
 import * as chalk from 'chalk';
 
 import { CommandMetadata } from '../../definitions';