+++ packages/@ionic/cli/src/commands/repair.ts
@@ -1,5 +1,5 @@
-import { prettyPath } from '@ionic/cli-framework/utils/format';
 import { pathExists, remove, unlink } from '@ionic/utils-fs';
+import { prettyPath } from '@ionic/utils-terminal';
 import * as path from 'path';
 
 import { CommandInstanceInfo, CommandLineInputs, CommandLineOptions, CommandMetadata, IProject, ProjectIntegration } from '../definitions';