+++ packages/@ionic/cli/src/constants.ts
@@ -1,3 +1,4 @@
+import * as chalk from 'chalk';
 import * as lodash from 'lodash';
 import * as path from 'path';
 
@@ -9,3 +10,5 @@ export const PROJECT_FILE = process.env['IONIC_CONFIG_FILE'] ?? 'ionic.config.js
 export const PROJECT_TYPES: ProjectType[] = ['angular', 'react', 'vue', 'ionic-angular', 'ionic1', 'custom'];
 export const LEGACY_PROJECT_TYPES: ProjectType[] = ['ionic-angular', 'ionic1'];
 export const MODERN_PROJECT_TYPES: ProjectType[] = lodash.difference(PROJECT_TYPES, LEGACY_PROJECT_TYPES);
+
+export const COLUMNAR_OPTIONS = { hsep: chalk.dim('-'), vsep: chalk.dim('|') };