+++ packages/@ionic/cli/src/lib/index.ts
@@ -1,7 +1,6 @@
 import { LOGGER_LEVELS } from '@ionic/cli-framework-output';
 import { createPromptModule } from '@ionic/cli-framework-prompts';
-import { prettyPath } from '@ionic/cli-framework/utils/format';
-import { TERMINAL_INFO } from '@ionic/utils-terminal';
+import { TERMINAL_INFO, prettyPath } from '@ionic/utils-terminal';
 import * as Debug from 'debug';
 import * as path from 'path';