+++ packages/@ionic/cli/src/lib/project/angular/serve.ts
@@ -1,7 +1,7 @@
 import { MetadataGroup, ParsedArgs, unparseArgs } from '@ionic/cli-framework';
-import { stripAnsi } from '@ionic/cli-framework-output';
 import { str2num } from '@ionic/cli-framework/utils/string';
 import { findClosestOpenPort } from '@ionic/utils-network';
+import { stripAnsi } from '@ionic/utils-terminal';
 import * as chalk from 'chalk';
 
 import { AngularServeOptions, CommandLineInputs, CommandLineOptions, CommandMetadata, ServeDetails } from '../../../definitions';