+++ packages/@ionic/utils-terminal/src/__tests__/format.ts
@@ -1,11 +1,45 @@
-import { stripAnsi } from '@ionic/cli-framework-output';
-
 import * as os from 'os';
 import * as path from 'path';
 
-describe('@ionic/cli-framework', () => {
+describe('@ionic/utils-terminal', () => {
 
-  describe('utils/format', () => {
+  describe('format', () => {
+
+    describe('wordWrap', () => {
+
+      const { wordWrap } = require('../format');
+
+      it('should do nothing to empty string', () => {
+        const result = wordWrap('', {});
+        expect(result).toEqual('');
+      });
+
+      it('should do nothing to newline', () => {
+        const result = wordWrap('\n', {});
+        expect(result).toEqual('\n');
+      });
+
+      it('should wrap words', () => {
+        const result = wordWrap('hello world', { width: 5 });
+        expect(result).toEqual('hello\nworld');
+      });
+
+      it('should wrap words with proper indentation', () => {
+        const result = wordWrap('hello world', { width: 5, indentation: 5 });
+        expect(result).toEqual('hello\n     world');
+      });
+
+      it('should wrap many words', () => {
+        const result = wordWrap(`I'm sorry Dave, I'm afraid I can't do that.`, { width: 10 });
+        expect(result).toEqual(`I'm sorry\nDave, I'm\nafraid I\ncan't do\nthat.`);
+      });
+
+      it('should wrap a command and append a character', () => {
+        const result = wordWrap(`git commit -m 'Initial commit' --no-gpg-sign`, { width: 40, indentation: 4, append: ' \\' });
+        expect(result).toEqual(`git commit -m 'Initial commit' \\\n    --no-gpg-sign`);
+      });
+
+    });
 
     describe('prettyPath', () => {
 
@@ -239,12 +273,12 @@ describe('@ionic/cli-framework', () => {
 
       it('should work without column headers', async () => {
         const result = columnar([['a', 'b', 'c']], {});
-        expect(stripAnsi(result)).toEqual('a | b | c');
+        expect(result).toEqual('a | b | c');
       });
 
       it('should work for multiple rows', async () => {
         const result = columnar([['cat', 'dog'], ['spongebob', 'squarepants'], ['hey', 'arnold']], { headers: ['col1', 'col2'] });
-        expect(stripAnsi(result)).toEqual(`
+        expect(result).toEqual(`
 col1      | col2
 -----------------------
 cat       | dog
@@ -255,7 +289,7 @@ hey       | arnold
 
       it('should size rows correctly when cells are longer', async () => {
         const result = columnar([['foo', 'bar', 'baz']], { headers: ['a', 'b', 'c'] });
-        expect(stripAnsi(result)).toEqual(`
+        expect(result).toEqual(`
 a   | b   | c
 ---------------
 foo | bar | baz
@@ -264,7 +298,7 @@ foo | bar | baz
 
       it('should size rows correctly when headers are longer', async () => {
         const result = columnar([['a', 'b', 'c']], { headers: ['foo', 'bar', 'baz'] });
-        expect(stripAnsi(result)).toEqual(`
+        expect(result).toEqual(`
 foo | bar | baz
 ---------------
 a   | b   | c
@@ -273,7 +307,7 @@ a   | b   | c
 
       it('should work with custom separators', async () => {
         const result = columnar([['foo', 'bar', 'baz']], { hsep: '=', vsep: '/\\', headers: ['a', 'b', 'c'] });
-        expect(stripAnsi(result)).toEqual(`
+        expect(result).toEqual(`
 a   /\\ b   /\\ c
 =================
 foo /\\ bar /\\ baz
@@ -282,7 +316,7 @@ foo /\\ bar /\\ baz
 
       it('should have appropriate spacing for cells with newlines', async () => {
         const result = columnar([['mr\n& mr', 'cat', 'dog,\nthe third'], ['mr', 'spongebob', 'squarepants'], ['', 'the', 'brain']], {});
-        expect(stripAnsi(result)).toEqual(`
+        expect(result).toEqual(`
 mr   | cat       | dog,
 & mr |           | the third
 mr   | spongebob | squarepants
@@ -292,7 +326,7 @@ mr   | spongebob | squarepants
 
       it('should separate columns and rows with only one space with empty string separators', async () => {
         const result = columnar([['a', 'b', 'c'], ['1', '2', '3']], { vsep: '', hsep: '', headers: ['x', 'x', 'x'] });
-        expect(stripAnsi(result)).toEqual(`
+        expect(result).toEqual(`
 x x x
 a b c
 1 2 3